refactor: XXXStatusの名前をもつenumをすべてapplication.enum直下に移動
