+++ app/src/main/java/com/punyo/nitechvacancyviewer/application/enums/AuthResultStatus.kt
@@ -1,0 +1,9 @@
+package com.punyo.nitechvacancyviewer.application.enums
+
+enum class AuthResultStatus {
+    SUCCESS,
+    INVALID_CREDENTIALS,
+    NETWORK_ERROR,
+    CREDENTIALS_NOT_FOUND,
+    UNKNOWN_ERROR,
+}
