+++ app/src/main/java/com/punyo/nitechvacancyviewer/application/enums/RoomVacancyStatus.kt
@@ -1,0 +1,6 @@
+package com.punyo.nitechvacancyviewer.application.enums
+
+enum class RoomVacancyStatus {
+    VACANT,
+    OCCUPY,
+}
